Moved video trim biz logic from video options panel to provider
